refactoring for future work
