+++ backend/service/departmentService.go
@@ -2,8 +2,8 @@ package service
 
 import (
 	"context"
-	"github.com/rojoherrero/quality-accounts/app/model"
-	"github.com/rojoherrero/quality-accounts/app/repository"
+	"github.com/rojoherrero/quality-accounts/backend/model"
+	"github.com/rojoherrero/quality-accounts/backend/repository"
 )
 
 //go:generate mockgen -source=$GOFILE -destination=../mock/mock_$GOFILE -package=mock
