+++ backend/service/roleService.go
@@ -5,8 +5,8 @@ package service
 import (
 	"context"
 
-	"github.com/rojoherrero/quality-accounts/app/model"
-	"github.com/rojoherrero/quality-accounts/app/repository"
+	"github.com/rojoherrero/quality-accounts/backend/model"
+	"github.com/rojoherrero/quality-accounts/backend/repository"
 )
 
 type (
